[phase 4] usage: merge Usage + Budgets into one tabbed page
